requirements.txt
certifi==2019.6.16
chardet==3.0.4
idna==2.8
Jinja2==2.10.1
MarkupSafe==1.1.1
Pillow==6.1.0
requests==2.22.0
urllib3==1.25.3
vaderSentiment==3.2.1
